requirements.txt
pillow
numpy
opencv-python
pytesseract
scipy
pytest
pytest-cov